wikignome/requirements.txt
fastapi
uvicorn
slack-sdk
groq
python-dotenv
markdown
jinja2
gTTS